🚚 Rename endpoints
ref : f431c3511e09275000b576114990184e4964d5dc
