+++ src/main/java/com/dodatabase/demo/controller/HomeController.java
@@ -10,6 +10,6 @@ public class HomeController {
 
   @GetMapping("")
   public String home() {
-    return "index";
+    return "html/index";
   }
 }
