+++ src/main/java/com/dodatabase/demo/service/WishListService.java
@@ -2,6 +2,7 @@ package com.dodatabase.demo.service;
 
 import com.dodatabase.demo.domain.movie.Movie;
 import com.dodatabase.demo.repository.MovieRepository;
+
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
